xml compiled with cxf

+++ core/src/test/java/org/metachart/factory/xml/chart/high/renderer/CliTimeSeriesChartRenderer.java
@@ -61,7 +61,7 @@ public class CliTimeSeriesChartRenderer
 		{
 			Data data = new Data();
 			data.setRecord(DateUtil.toXmlGc(LocalDate.of(2010, 1, i)));
-			data.setY(rnd.nextInt(i));
+			data.setY(Integer.valueOf(rnd.nextInt(i)).doubleValue());
 			if(rnd.nextInt(100)<70){dataSet.getData().add(data);}
 		}
 		return dataSet;
